text2image_iep/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
fastapi
uvicorn[standard]
transformers
qdrant-client
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pydantic
pillow
python-dotenv

